+++ benchmark/byte_queue_view_benchmark.cpp
@@ -1,4 +1,4 @@
-#include "galay-utils/buffer/byte_queue_view.hpp"
+#include "galay-utils/tool/byte_queue_view.hpp"
 
 #include <chrono>
 #include <cstdint>